backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
snowflake-connector-python==3.6.0
fastmcp==0.4.2
mcp==1.0.0
python-dotenv==1.0.0
aiohttp==3.9.3
starlette==0.35.1
pydantic-settings==2.1.0
